# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Management System API",
    "description": "API documentation for the Employee Management System, including secure endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/employees/{id}": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "getEmployeeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/attendance/punch-in": {
      "post": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "punchIn",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OFFICE",
                "WFH",
                "FIELD"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wfh/employee/{employeeId}": {
      "get": {
        "tags": [
          "wfh-request-controller"
        ],
        "operationId": "getWfhRequestsByEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WfhRequestDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews/employee/{employeeId}": {
      "get": {
        "tags": [
          "performance-review-controller"
        ],
        "operationId": "getEmployeeReviews",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerformanceReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/leaves/employee/{employeeId}": {
      "get": {
        "tags": [
          "leave-request-controller"
        ],
        "operationId": "getLeaveRequestsByEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaveRequestDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/attendance/hours": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "getTotalHours",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/api/attendance/employee/{employeeId}": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "getEmployeeAttendance",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/attendance/date/{date}": {
      "get": {
        "tags": [
          "attendance-controller"
        ],
        "operationId": "getAllAttendanceByDate",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WfhRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "reason": {
            "type": "string"
          },
          "approvedById": {
            "type": "string",
            "format": "uuid"
          },
          "approvedByName": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LeaveRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "leaveType": {
            "type": "string",
            "enum": [
              "SICK",
              "CASUAL",
              "ANNUAL",
              "UNPAID"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "approvedById": {
            "type": "string",
            "format": "uuid"
          },
          "approvedByName": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "endDate",
          "leaveType",
          "reason",
          "startDate"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 40,
            "minLength": 6
          },
          "phone": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "departmentCode": {
            "type": "string"
          },
          "managerId": {
            "type": "string",
            "format": "uuid"
          },
          "salary": {
            "type": "number"
          },
          "dateOfJoining": {
            "type": "string",
            "format": "date"
          },
          "role": {
            "type": "string",
            "enum": [
              "ROLE_ADMIN",
              "ROLE_MANAGER",
              "ROLE_EMPLOYEE"
            ]
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ]
      },
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "departmentId": {
            "type": "string",
            "format": "uuid"
          },
          "departmentCode": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "salary": {
            "type": "number"
          },
          "dateOfJoining": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "TERMINATED",
              "ON_LEAVE"
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "ROLE_ADMIN",
              "ROLE_MANAGER",
              "ROLE_EMPLOYEE"
            ]
          },
          "managerId": {
            "type": "string",
            "format": "uuid"
          },
          "managerName": {
            "type": "string"
          },
          "managerEmail": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PerformanceReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeName": {
            "type": "string"
          },
          "reviewerId": {
            "type": "string",
            "format": "uuid"
          },
          "reviewerName": {
            "type": "string"
          },
          "reviewDate": {
            "type": "string",
            "format": "date"
          },
          "feedback": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "employeeId",
          "feedback",
          "rating"
        ]
      },
...
      "AttendanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "employeeName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "clockIn": {
            "type": "string",
            "format": "date-time"
          },
          "clockOut": {
            "type": "string",
            "format": "date-time"
          },
          "workLocation": {
            "type": "string",
            "enum": [
              "OFFICE",
              "WFH",
              "FIELD"
            ]
          },
          "workHours": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}